pyproject.toml
[project]
name = "billing-services"
version = "0.1.0"
description = ""
authors = [
    {name = "Thomas Wolter",email = "thwolter@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "pydantic (>=2.11.7,<3.0.0)",
    "openmeter (>=1.0.0b188,<2.0.0)",
    "azure-core (>=1.34.0,<2.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "aiobreaker (>=1.2.0,<2.0.0)",
    "aiohttp (>=3.12.13,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
isort = "^6.0.1"
ruff = "^0.12.0"
mypy = "^1.16.1"
mkdocs = "^1.6.1"
pytest-asyncio = "^1.0.0"

